refactor: move addr masking to utils directory
