+++ src/store/wallet-store.tsx
@@ -5,7 +5,7 @@ import { getPreferredWeb3ProviderAsync } from "@/utils/ethereum-provider";
 import { getAllWalletProviders } from "@/utils/ethereum-provider";
 import { createWalletError, WalletNotFoundError } from "@/utils/wallet-errors";
 import { fetchUserTransactions, mapApiTransactionsToAppFormat, fetchFeeEstimation } from "@/services/api";
-import { maskAddress } from "@/lib/utils";
+import { maskAddress } from "@/utils";
 import { getMetaMaskProvider } from "@/utils/ethereum-provider";
 
 // Create events for wallet state changes
